+++ packages/expo-router/src/fork/__tests__/getPathFromState.test.web.ts
@@ -1,4 +1,4 @@
-import getPathFromState from '../getPathFromState';
+import { getPathFromState } from '../getPathFromState';
 
 describe('hash support', () => {
   it('appends hash to the path', () => {