+++ packages/expo-router/src/testing-library/index.tsx
@@ -7,7 +7,7 @@ import React from 'react';
 
 import { MockContextConfig, getMockConfig, getMockContext } from './mock-config';
 import { ExpoRoot } from '../ExpoRoot';
-import getPathFromState from '../fork/getPathFromState';
+import { getPathFromState } from '../fork/getPathFromState';
 import { ExpoLinkingOptions } from '../getLinkingConfig';
 import { store } from '../global-state/router-store';
 import { router } from '../imperative-api';