+++ packages/expo-router/src/ui/Tabs.tsx
@@ -1,5 +1,4 @@
 import {
-  DefaultNavigatorOptions,
   LinkingContext,
   ParamListBase,
   TabActionHelpers,
@@ -18,6 +17,7 @@ import {
 import { StyleSheet, ViewProps, View } from 'react-native';
 
 import {
+  ExpoTabsProps,
   ExpoTabsScreenOptions,
   TabNavigationEventMap,
   TabTriggerMapContext,
@@ -41,15 +41,7 @@ export * from './TabList';
 export * from './TabSlot';
 export * from './TabTrigger';
 
-export type UseTabsOptions = Omit<
-  DefaultNavigatorOptions<
-    ParamListBase,
-    TabNavigationState<any>,
-    ExpoTabsScreenOptions,
-    TabNavigationEventMap
-  >,
-  'children'
-> &
+export type UseTabsOptions = Omit<ExpoTabsProps, 'children'> &
   Omit<ExpoTabRouterOptions, 'initialRouteName' | 'triggerMap'>;
 
 export type TabsProps = ViewProps & {