+++ packages/expo-router/src/views/Sitemap.tsx
@@ -30,7 +30,6 @@ export function getNavOptions(): NativeStackNavigationOptions {
     title: 'sitemap',
     headerShown: false,
     presentation: 'modal',
-    animation: 'default',
     headerLargeTitle: false,
     headerTitleStyle: {
       color: 'white',