+++ packages/expo-router/types/react-navigation.d.ts
@@ -1,0 +1,2 @@
+declare module '@react-navigation/native/lib/commonjs/createMemoryHistory';
+declare module '@react-navigation/native/lib/commonjs/UnhandledLinkingContext';